// src/models/shopping-item/shopping-item-interface.ts
export interface ShoppingItem{
      
      $key?: string;
      itemName: string;
      valor: number;
      myDate: Date;
      tipoDivida: string;
      qtdParcela: number;
      valorParcela:number;
      tipoPagamento: string;
      histValor: number;
      histValorParcela:number;
      pega:number;
      

}